+++ internal/pkg/warewulfd/warewulfd.go
@@ -10,6 +10,8 @@ import (
 	"syscall"
 
 	warewulfconf "github.com/warewulf/warewulf/internal/pkg/config"
+	"github.com/warewulf/warewulf/internal/pkg/util"
+	"github.com/warewulf/warewulf/internal/pkg/warewulfd/api"
 	"github.com/warewulf/warewulf/internal/pkg/wwlog"
 )
 
@@ -31,6 +33,20 @@ Filter out the '//'
 func (h *slashFix) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	r.URL.Path = strings.Replace(r.URL.Path, "//", "/", -1)
 	h.mux.ServeHTTP(w, r)
+}
+
+func defaultHandler() *slashFix {
+	var wwHandler http.ServeMux
+	wwHandler.HandleFunc("/provision/", ProvisionSend)
+	wwHandler.HandleFunc("/ipxe/", ProvisionSend)
+	wwHandler.HandleFunc("/efiboot/", ProvisionSend)
+	wwHandler.HandleFunc("/kernel/", ProvisionSend)
+	wwHandler.HandleFunc("/container/", ProvisionSend)
+	wwHandler.HandleFunc("/overlay-system/", ProvisionSend)
+	wwHandler.HandleFunc("/overlay-runtime/", ProvisionSend)
+	wwHandler.HandleFunc("/overlay-file/", OverlaySend)
+	wwHandler.HandleFunc("/status", StatusSend)
+	return &slashFix{&wwHandler}
 }
 
 func RunServer() error {
@@ -65,22 +81,28 @@ func RunServer() error {
 	if err != nil {
 		wwlog.Warn("couldn't copy default shim: %s", err)
 	}
-	var wwHandler http.ServeMux
-	wwHandler.HandleFunc("/provision/", ProvisionSend)
-	wwHandler.HandleFunc("/ipxe/", ProvisionSend)
-	wwHandler.HandleFunc("/efiboot/", ProvisionSend)
-	wwHandler.HandleFunc("/kernel/", ProvisionSend)
-	wwHandler.HandleFunc("/image/", ProvisionSend)
-	wwHandler.HandleFunc("/overlay-system/", ProvisionSend)
-	wwHandler.HandleFunc("/overlay-runtime/", ProvisionSend)
-	wwHandler.HandleFunc("/overlay-file/", OverlaySend)
-	wwHandler.HandleFunc("/status", StatusSend)
 
 	conf := warewulfconf.Get()
-
 	daemonPort := conf.Warewulf.Port
-	err = http.ListenAndServe(":"+strconv.Itoa(daemonPort), &slashFix{&wwHandler})
 
+	auth := warewulfconf.NewAuthentication()
+	if util.IsFile(conf.Paths.AuthenticationConf()) {
+		err = auth.Read(conf.Paths.AuthenticationConf())
+		if err != nil {
+			wwlog.Warn("%w\n", err)
+		}
+	}
+
+	apiHandler := api.Handler(auth)
+	defaultHandler := defaultHandler()
+	dispatchHandler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		if strings.HasPrefix(r.URL.Path, "/api") && conf.API != nil && conf.API.Enabled() {
+			apiHandler.ServeHTTP(w, r)
+		} else {
+			defaultHandler.ServeHTTP(w, r)
+		}
+	})
+	err = http.ListenAndServe(":"+strconv.Itoa(daemonPort), dispatchHandler)
 	if err != nil {
 		return fmt.Errorf("could not start listening service: %w", err)
 	}